requirements.txt
Flask
Waitress
pyyaml
subprocess32